fix named problem and superfluous word
